Use Chrome Extension URL to inject the page script. This would allow us better debug the content script. However this breaks the i18n feature inside the content script.
Dynamically search for the `subscribe` function on the `player` object. This makes the script compatible
to futrue updates of YouTube internal APIs on this specific function.

TODO: fix the i18n feature inside the content script. We need to wrap the `window.postMessage` async
call and make it synchronized using busy waiting. Then we can implement a `gettext` like function to do
the i18n job.